Moved indy-data-types into anoncreds module data_types
