+++ Cargo.toml
@@ -2,7 +2,6 @@
 
 members = [
     "anoncreds",
-    "indy-data-types",
     "indy-utils",
     "indy-wql"
 ]
