+++ anoncreds/src/data_types/anoncreds/cred_def.rs
@@ -1,12 +1,11 @@
-use crate::identifiers::cred_def::CredentialDefinitionId;
-use crate::identifiers::schema::SchemaId;
-use crate::utils::Qualifiable;
-use crate::{ConversionError, Validatable, ValidationError};
+use crate::data_types::identifiers::cred_def::CredentialDefinitionId;
+use crate::data_types::identifiers::schema::SchemaId;
+use crate::data_types::utils::Qualifiable;
+use crate::data_types::{ConversionError, Validatable, ValidationError};
 
 pub const CL_SIGNATURE_TYPE: &str = "CL";
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub enum SignatureType {
     CL,
 }
@@ -26,18 +25,17 @@ impl SignatureType {
     }
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct CredentialDefinitionData {
     pub primary: ursa_cl!(CredentialPrimaryPublicKey),
-    #[cfg_attr(feature = "serde", serde(skip_serializing_if = "Option::is_none"))]
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub revocation: Option<ursa_cl!(CredentialRevocationPublicKey)>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize), serde(tag = "ver"))]
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(tag = "ver")]
 pub enum CredentialDefinition {
-    #[cfg_attr(feature = "serde", serde(rename = "1.0"))]
+    #[serde(rename = "1.0")]
     CredentialDefinitionV1(CredentialDefinitionV1),
 }
 
@@ -71,22 +69,17 @@ impl Validatable for CredentialDefinition {
     }
 }
 
-#[derive(Debug)]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(rename_all = "camelCase")
-)]
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct CredentialDefinitionV1 {
     pub id: CredentialDefinitionId,
     pub schema_id: SchemaId,
-    #[cfg_attr(feature = "serde", serde(rename = "type"))]
+    #[serde(rename = "type")]
     pub signature_type: SignatureType,
     pub tag: String,
     pub value: CredentialDefinitionData,
 }
 
-#[cfg(any(feature = "cl", feature = "cl_native"))]
 impl CredentialDefinitionV1 {
     pub fn get_public_key(&self) -> Result<crate::ursa::cl::CredentialPublicKey, ConversionError> {
         let key = crate::ursa::cl::CredentialPublicKey::build_from_parts(
@@ -105,31 +98,21 @@ impl Validatable for CredentialDefinitionV1 {
     }
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct CredentialDefinitionPrivate {
     pub value: ursa_cl!(CredentialPrivateKey),
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize), serde(transparent))]
+#[derive(Debug, Deserialize, Serialize)]
+#[serde(transparent)]
 pub struct CredentialKeyCorrectnessProof {
     pub value: ursa_cl!(CredentialKeyCorrectnessProof),
 }
 
 impl CredentialKeyCorrectnessProof {
     pub fn try_clone(&self) -> Result<Self, ConversionError> {
-        #[cfg(any(feature = "cl", feature = "cl_native"))]
-        {
-            Ok(Self {
-                value: self.value.try_clone().map_err(|e| e.to_string())?,
-            })
-        }
-        #[cfg(not(any(feature = "cl", feature = "cl_native")))]
-        {
-            Ok(Self {
-                value: self.value.clone(),
-            })
-        }
+        Ok(Self {
+            value: self.value.try_clone().map_err(|e| e.to_string())?,
+        })
     }
 }
