+++ anoncreds/src/data_types/anoncreds/macros.rs
@@ -1,0 +1,5 @@
+macro_rules! ursa_cl {
+    ($ident:ident) => {
+        $crate::ursa::cl::$ident
+    };
+}
