+++ anoncreds/src/data_types/anoncreds/presentation.rs
@@ -1,30 +1,28 @@
 use std::collections::HashMap;
 
-use crate::identifiers::cred_def::CredentialDefinitionId;
-use crate::identifiers::rev_reg::RevocationRegistryId;
-use crate::identifiers::schema::SchemaId;
-use crate::Validatable;
+use crate::data_types::identifiers::cred_def::CredentialDefinitionId;
+use crate::data_types::identifiers::rev_reg::RevocationRegistryId;
+use crate::data_types::identifiers::schema::SchemaId;
+use crate::data_types::Validatable;
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Presentation {
     pub proof: ursa_cl!(Proof),
     pub requested_proof: RequestedProof,
     pub identifiers: Vec<Identifier>,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
 pub struct RequestedProof {
     pub revealed_attrs: HashMap<String, RevealedAttributeInfo>,
-    #[cfg_attr(feature = "serde", serde(skip_serializing_if = "HashMap::is_empty"))]
-    #[cfg_attr(feature = "serde", serde(default))]
+    #[serde(skip_serializing_if = "HashMap::is_empty")]
+    #[serde(default)]
     pub revealed_attr_groups: HashMap<String, RevealedAttributeGroupInfo>,
-    #[cfg_attr(feature = "serde", serde(default))]
+    #[serde(default)]
     pub self_attested_attrs: HashMap<String, String>,
-    #[cfg_attr(feature = "serde", serde(default))]
+    #[serde(default)]
     pub unrevealed_attrs: HashMap<String, SubProofReferent>,
-    #[cfg_attr(feature = "serde", serde(default))]
+    #[serde(default)]
     pub predicates: HashMap<String, SubProofReferent>,
 }
 
@@ -40,36 +38,31 @@ impl Default for RequestedProof {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
 pub struct SubProofReferent {
     pub sub_proof_index: u32,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
 pub struct RevealedAttributeInfo {
     pub sub_proof_index: u32,
     pub raw: String,
     pub encoded: String,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
 pub struct RevealedAttributeGroupInfo {
     pub sub_proof_index: u32,
     pub values: HashMap<String /* attribute name */, AttributeValue>,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
 pub struct AttributeValue {
     pub raw: String,
     pub encoded: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Deserialize, Serialize)]
 pub struct Identifier {
     pub schema_id: SchemaId,
     pub cred_def_id: CredentialDefinitionId,
@@ -81,10 +74,8 @@ impl Validatable for Presentation {}
 
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "serde")]
     use super::*;
 
-    #[cfg(feature = "serde")]
     #[test]
     fn deserialize_requested_proof_with_empty_revealed_attr_groups() {
         let mut req_proof_old: RequestedProof = Default::default();
