+++ anoncreds/src/data_types/identifiers/cred_def.rs
@@ -1,6 +1,6 @@
 use super::schema::SchemaId;
-use crate::utils::{qualifiable, Qualifiable};
-use crate::{Validatable, ValidationError};
+use crate::data_types::utils::{qualifiable, Qualifiable};
+use crate::data_types::{Validatable, ValidationError};
 use indy_utils::did::DidValue;
 use indy_utils::qualifiable_type;
