+++ anoncreds/src/data_types/identifiers/schema.rs
@@ -1,5 +1,5 @@
-use crate::utils::{qualifiable, Qualifiable};
-use crate::{Validatable, ValidationError};
+use crate::data_types::utils::{qualifiable, Qualifiable};
+use crate::data_types::{Validatable, ValidationError};
 use indy_utils::did::DidValue;
 use indy_utils::qualifiable_type;
