+++ anoncreds/src/data_types/merkle_tree/mod.rs
@@ -342,7 +342,6 @@ mod tests {
         }
     }
 
-    #[cfg(feature = "serde")]
     #[test]
     fn serialize_works() {
         let strvals = vec!["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"];
