+++ anoncreds/src/services/verifier.rs
@@ -7,7 +7,7 @@ use super::helpers::*;
 use super::types::*;
 use crate::error::Result;
 use crate::ursa::cl::{verifier::Verifier as CryptoVerifier, CredentialPublicKey};
-use indy_data_types::anoncreds::{
+use crate::data_types::anoncreds::{
     nonce::Nonce,
     pres_request::{AttributeInfo, NonRevocedInterval, PredicateInfo, PresentationRequestPayload},
     presentation::{Identifier, RequestedProof, RevealedAttributeInfo},
