--- indy-data-types/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "indy-data-types"
-version = "0.5.1"
-authors = ["Hyperledger Indy Contributors <indy@lists.hyperledger.org>"]
-description = "Common data types for Hyperledger Indy (https://www.hyperledger.org/projects), which provides a distributed-ledger-based foundation for self-sovereign identity (https://sovrin.org)."
-edition = "2018"
-license = "Apache-2.0"
-readme = "../README.md"
-repository = "https://github.com/hyperledger/indy-shared-rs/"
-categories = ["authentication", "cryptography"]
-keywords = ["hyperledger", "indy", "ssi", "verifiable", "credentials"]
-
-[lib]
-name = "indy_data_types"
-path = "src/lib.rs"
-crate-type = ["rlib"]
-
-[features]
-default = ["merkle_tree", "serde_support"]
-cl = ["serde_support", "ursa", "ursa/cl", "ursa/serde"]
-cl_native = ["serde_support", "ursa", "ursa/cl_native", "ursa/serde"]
-merkle_tree = ["indy-utils/hash", "hex"]
-rich_schema = []
-serde_support = ["serde", "serde_json"]
-vendored = ["openssl", "openssl/vendored"]
-
-[dependencies]
-hex = { version = "0.4", optional = true }
-openssl = { version = "0.10", optional = true }
-once_cell = "1.9"
-regex = "1.3"
-serde = { version = "1.0", optional = true, features = ["derive"] }
-serde_json = { version = "1.0", optional = true, features = ["raw_value"] }
-ursa = { version = "=0.3.6", default-features = false, optional = true }
-zeroize = { version = "1.1", features = ["zeroize_derive"] }
-
-[dependencies.indy-utils]
-version = "0.5"
-path = "../indy-utils"
-default-features = false
-features = ["wql"]
-
-[dev-dependencies]
-hex = "0.4"
-serde_json = "1.0"
